Funcionalidades para Cerberus
- Movi funcionalidades do Sprint 2 para a Cerberus
- Corrigi detalhes da resposta HTTP das atividades
