+++ arquivos/atividades/3/7/Lab1/entrada
